+++ packages/test-factory/CHANGELOG.md
@@ -1,5 +1,15 @@
 # orchid-orm-test-factory
 
+## 0.3.18
+
+### Patch Changes
+
+- Updated dependencies
+  - orchid-core@0.1.7
+  - pqb@0.10.14
+  - orchid-orm@1.6.18
+  - orchid-orm-schema-to-zod@0.3.14
+
 ## 0.3.17
 
 ### Patch Changes